feat(samples): switch customer sales values to decimal and adjust queries

+++ sample/Atc.Kusto.Api.Sample/Contracts/CustomerSales.cs
@@ -3,5 +3,5 @@ namespace Atc.Kusto.Api.Sample.Contracts;
 public record CustomerSales(
     int CustomerKey,
     string CustomerName,
-    float SalesAmount,
-    float TotalCost);
+    decimal SalesAmount,
+    decimal TotalCost);
