+++ sample/Atc.Kusto.Sample/Contracts/CustomerSales.cs
@@ -3,5 +3,5 @@ namespace Atc.Kusto.Sample.Contracts;
 public record CustomerSales(
     int CustomerKey,
     string CustomerName,
-    float SalesAmount,
-    float TotalCost);
+    decimal SalesAmount,
+    decimal TotalCost);
